added more tests
